스케줄 표시 성공!!
스케줄 표시 성공!!

+++ app/src/main/java/com/goingbacking/goingbacking/Model/Event.kt
@@ -4,7 +4,7 @@ data class Event(
     var dest: String? = null,
     var date: String? = null,
     var start: Int? = null,
-    var start_t: Long? = null,
+    var start_t: Int? = null,
     var end: Int? = null,
-    var end_t: Long? = null,
+    var end_t: Int? = null,
 )
